# docs/swagger.yaml
definitions:
...
  common.KYCLevel:
    enum:
    - "0"
    - "1"
    - "2"
    - "3"
    type: string
    x-enum-varnames:
    - KYC_LEVEL_0
    - KYC_LEVEL_1
    - KYC_LEVEL_2
    - KYC_LEVEL_3
  common.KYCStatus:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - KYC_STATUS_PENDING
    - KYC_STATUS_PASS
    - KYC_STATUS_REJECT
    - KYC_STATUS_CREATED
    - KYC_STATUS_FAILED
  common.Language:
    enum:
    - zh
    - zh_hans
    - zh_hant
    - zh_hk
    - zh_sg
    - en
    - en_us
    - en_gb
    - en_au
    - en_ca
    - de
    - de_de
    - de_at
    - de_ch
    - fr
    - fr_fr
    - fr_ca
    - es
    - es_es
    - es_mx
    - it
    - pt
    - pt_br
    - nl
    - ru
    - pl
    - tr
    - ja
    - ko
    - th
    - vi
    - hi
    - ar
    - af
    - bg
    - ca
    - cs
    - da
    - el
    - fi
    - he
    - hu
    - id
    - "no"
    - ro
    - sk
    - sv
    - uk
    type: string
    x-enum-varnames:
    - LANGUAGE_CHINESE
    - LANGUAGE_SIMPLIFIED_CHINESE
    - LANGUAGE_TRADITIONAL_CHINESE_TW
    - LANGUAGE_TRADITIONAL_CHINESE_HK
    - LANGUAGE_SIMPLIFIED_CHINESE_SG
    - LANGUAGE_ENGLISH
    - LANGUAGE_ENGLISH_US
    - LANGUAGE_ENGLISH_GB
    - LANGUAGE_ENGLISH_AU
    - LANGUAGE_ENGLISH_CA
    - LANGUAGE_GERMAN
    - LANGUAGE_GERMAN_DE
    - LANGUAGE_GERMAN_AT
    - LANGUAGE_GERMAN_CH
    - LANGUAGE_FRENCH
    - LANGUAGE_FRENCH_FR
    - LANGUAGE_FRENCH_CA
    - LANGUAGE_SPANISH
    - LANGUAGE_SPANISH_ES
    - LANGUAGE_SPANISH_MX
    - LANGUAGE_ITALIAN
    - LANGUAGE_PORTUGUESE
    - LANGUAGE_PORTUGUESE_BR
    - LANGUAGE_DUTCH
    - LANGUAGE_RUSSIAN
    - LANGUAGE_POLISH
    - LANGUAGE_TURKISH
    - LANGUAGE_JAPANESE
    - LANGUAGE_KOREAN
    - LANGUAGE_THAI
    - LANGUAGE_VIETNAMESE
    - LANGUAGE_HINDI
    - LANGUAGE_ARABIC
    - LANGUAGE_AFRIKAANS
    - LANGUAGE_BULGARIAN
    - LANGUAGE_CATALAN
    - LANGUAGE_CZECH
    - LANGUAGE_DANISH
    - LANGUAGE_GREEK
    - LANGUAGE_FINNISH
    - LANGUAGE_HEBREW
    - LANGUAGE_HUNGARIAN
    - LANGUAGE_INDONESIAN
    - LANGUAGE_NORWEGIAN
    - LANGUAGE_ROMANIAN
    - LANGUAGE_SLOVAK
    - LANGUAGE_SWEDISH
    - LANGUAGE_UKRAINIAN
...
  common.Role:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    type: integer
    x-enum-varnames:
    - ROLE_GUEST
    - ROLE_USER
    - ROLE_ADMIN
    - ROLE_MERCHANT
    - ROLE_MERCHANT_ADMIN
    - ROLE_MERCHANT_USER
    - ROLE_ADMIN_AGENT_USER
    - ROLE_DAPP_USER
    - ROLE_SYSTEM
...
  common.UserBlockStatus:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - USER_BLOCK_STATUS_NORMAL
    - USER_BLOCK_STATUS_BLOCK
...
  shared-modules_entities.CreateWalletForm:
    properties:
      categoryId:
        example: "0"
        type: string
      currency:
        type: string
    type: object
...
  shared-modules_entities.DeleteAccountForm:
    properties:
      pinCode:
        type: string
    required:
    - pinCode
    type: object
...
  shared-modules_entities.DeviceInfo:
    properties:
      deviceCount:
        type: integer
      deviceID:
        type: string
      deviceName:
        type: string
      devicePasskey:
        type: boolean
      hasPasskey:
        type: boolean
      isMaster:
        type: boolean
      platform:
        type: string
    type: object
...
  shared-modules_entities.ForgotPinCodeForm:
    properties:
      deviceID:
        type: string
      id:
        type: string
      newPinCode:
        type: string
      otpCode:
        type: string
      rawId:
        type: string
      response:
        $ref: '#/definitions/shared-modules_entities.VerifyResponseType'
      type:
        type: string
    required:
    - newPinCode
    type: object
...
  shared-modules_entities.GetInfoForm:
    type: object
  shared-modules_entities.GetInfoVO:
    properties:
      LastLoginAt:
        example: "0"
        type: string
      atmToggle:
        type: string
      auto3ds:
        type: string
      autoTopUp:
        type: string
      blockStatus:
        $ref: '#/definitions/common.UserBlockStatus'
      channel:
        type: string
      countryCode:
        description: 可能為空
        type: integer
      createdAt:
        example: "0"
        type: string
      device:
        $ref: '#/definitions/shared-modules_entities.DeviceInfo'
      ePoint:
        type: number
      ePointCardId:
        example: "0"
        type: string
      ePointLevel:
        type: string
      email:
        type: string
      finishedIdentity:
        type: string
      gender:
        type: string
      groupIDs:
        items:
          type: integer
        type: array
      id:
        example: "0"
        type: string
      identity:
        $ref: '#/definitions/shared-modules_entities.IdentityVO'
      kycLevel:
        $ref: '#/definitions/common.KYCLevel'
      kycStatus:
        $ref: '#/definitions/common.KYCStatus'
      language:
        $ref: '#/definitions/common.Language'
      phoneNumber:
        description: 可能為空
        type: string
      promotionCode:
        type: string
      promotionLink:
        type: string
      referrerCode:
        type: string
      role:
        $ref: '#/definitions/common.Role'
      updatedAt:
        example: "0"
        type: string
    type: object
...
  shared-modules_entities.IdentityVO:
    properties:
      ValidatedAt:
        type: string
      addressLine1:
        type: string
      addressLine2:
        type: string
      city:
        type: string
      country:
        type: string
      dayOfBornAt:
        type: string
      dob:
        type: string
      failReason:
        type: string
      firstName:
        type: string
      idDocumentNumber:
        type: string
      idDocumentType:
        type: integer
      lastName:
        type: string
      nationCode:
        type: string
      postalCode:
        type: string
      state:
        type: string
      status:
        $ref: '#/definitions/common.KYCStatus'
    type: object
...
  shared-modules_entities.ListWalletsForm:
    properties:
      categoryID:
        type: integer
      id:
        type: integer
    type: object
...
  shared-modules_entities.ResetPinCodeForm:
    properties:
      deviceID:
        type: string
      id:
        type: string
      newPinCode:
        type: string
      oldPinCode:
        type: string
      rawId:
        type: string
      response:
        $ref: '#/definitions/shared-modules_entities.VerifyResponseType'
      type:
        type: string
    required:
    - newPinCode
    type: object
...
  shared-modules_entities.SaveIdentityVO:
    type: object
  shared-modules_entities.SaveLanguageForm:
    properties:
      language:
        $ref: '#/definitions/common.Language'
    type: object
  shared-modules_entities.SavePhoneNumberForm:
    properties:
      countryCode:
        type: string
      phoneNumber:
        type: string
    type: object
...
  shared-modules_entities.SetPinCodeForm:
    properties:
      pinCode:
        type: string
    required:
    - pinCode
    type: object
  shared-modules_entities.SetPinCodeVO:
    type: object
...
  shared-modules_entities.VerifyResponseType:
    properties:
      authenticatorData:
        type: string
      clientDataJSON:
        type: string
      signature:
        type: string
      userHandle:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /web/user/deleteAccount:
    post:
      consumes:
      - application/json
      description: 刪除指定用戶的帳戶
      parameters:
      - description: Token
        in: header
        name: X-Token
        required: true
        type: string
      - description: Delete Account Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.DeleteAccountForm'
      produces:
      - application/json
      responses: {}
      summary: 刪除用戶帳戶
      tags:
      - web/user
  /web/user/forgotPinCode:
    post:
      description: forget user PIN code.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ForgotPinCodeForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            type: string
      tags:
      - web/user
  /web/user/getInfo:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.GetInfoForm'
      - description: X-Token
        in: header
        name: X-Token
        required: true
        type: string
      - description: X-Extend
        in: header
        name: X-Extend
        type: string
      - description: X-Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.GetInfoVO'
      tags:
      - web/user
...
  /web/user/resetPinCode:
    post:
      description: reset user PIN code.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ResetPinCodeForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.SetPinCodeVO'
      tags:
      - web/user
...
  /web/user/saveLanguage:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.SaveLanguageForm'
      - description: X-Token
        in: header
        name: X-Token
        required: true
        type: string
      - description: X-Extend
        in: header
        name: X-Extend
        type: string
      responses: {}
      tags:
      - web/user
  /web/user/savePhoneNumber:
    post:
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.SavePhoneNumberForm'
      - description: X-Token
        in: header
        name: X-Token
        required: true
        type: string
      - description: X-Extend
        in: header
        name: X-Extend
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.SaveIdentityVO'
      tags:
      - web/user
...
  /web/user/setPinCode:
    post:
      description: Set user PIN code.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.SetPinCodeForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.SetPinCodeVO'
      tags:
      - web/user
...
  /web/wallet/apply:
    post:
      description: Create a new wallet for the user <br> 幣種列表： BTC ETH  USDT USDC
        DAI  WBTC  TRX  ADA  BCH  DOGE  LTC  XRP  SOL  BNB  ETC  MATIC
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.CreateWalletForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses: {}
      summary: Create a new wallet
      tags:
      - web/wallet
...
  /web/wallet/list:
    post:
      description: Get all wallets of the user
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ListWalletsForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses: {}
      summary: Get all wallets
      tags:
      - web/wallet
...
swagger: "2.0"